pyproject.toml
[project]
name = "signalforge"
version = "0.1.0"
description = "SignalForge - AI-powered trending news intelligence MCP server for filtering signal from noise"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.32.5,<3.0.0",
    "pytz>=2025.2,<2026.0",
    "PyYAML>=6.0.3,<7.0.0",
    "fastmcp>=2.12.0,<2.14.0",
    "websockets>=13.0,<14.0",
]

[project.scripts]
signalforge = "mcp_server.server:run_server"

[dependency-groups]
dev = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mcp_server"]
